Move the linear tree shap p2s and logic to their own files. The used logic move to the used p2s and most of the logic, which is unused move to the archive package.
also move get_unique_features_in_path and get_covers_vector to utils
