+++ woodelf/core/path_to_s_vectors/archive/lts_polynomial_multiplication.py
@@ -3,50 +3,11 @@ from __future__ import annotations
 import itertools
 import numpy as np
 
-
-def bits_matrix(x: np.ndarray, k: int) -> np.ndarray:
-    """
-    x: shape (n,), integers
-    returns: shape (k, n), rows are bits (k-1),...,1,0 (2^(k-1) down to LSB)
-    """
-    # ensure x is unsigned (np.uint) for fast bit ops
-    shifts = np.arange(k-1, -1, -1, dtype=np.uint8)[:, None]  # (5,1): 4,3,2,1,0
-    return ((x[None, :] >> shifts) & 1).astype(np.uint8)
-
-def neg_bits_matrix(x: np.ndarray, k: int) -> np.ndarray:
-    """
-    identical to bits_matrix(x,k).replace({0:1, 1:0})
-    """
-    # ensure x is unsigned (np.uint) for fast bit ops
-    shifts = np.arange(k-1, -1, -1, dtype=np.uint8)[:, None]  # (5,1): 4,3,2,1,0
-    return (((x[None, :] >> shifts) + 1) & 1).astype(np.uint8)
+from woodelf.core.utils import bits_matrix, neg_bits_matrix
 
 
-def linear_tree_shap_magic_for_banzhaf(
-        r: np.array, p: np.array, leaf_weight: float
-):
-    R_emptyset = np.prod(r) * leaf_weight
-    q_M = bits_matrix(p, len(r)) * (1/r.reshape(-1, 1))
-    sum_original_coefs = np.prod((1 + q_M), axis=0) * R_emptyset
-    constitutions_vectors = []
-    for i in range(len(r)):
-        M_f_i = sum_original_coefs * (1/(1+q_M[i]))
-        # Compute Banzhaf values using the constructed polynomial
-        game_theory_metric_vector = M_f_i / 2 ** (len(r) - 1)
-        constitutions_vectors.append(game_theory_metric_vector)
 
-    M = np.array(constitutions_vectors)  # Now M become a |n| columns and |r| rows matrix
-    return (M * (q_M - 1)).T.copy()
-
-
-############################################################################################################################################################
-#
-#                         The Recursive Approach (Without the Neighbor Leaf Trick)
-#
-############################################################################################################################################################
-
-
-def poly_mul_y_plus_q_inplace(P: np.ndarray, q: np.ndarray) -> None:
+def poly_mul_y_plus_q_inplace_archive(P: np.ndarray, q: np.ndarray) -> None:
     """
     In-place multiply polynomial P(y) by (y + q) for many columns at once.
 
@@ -64,31 +25,29 @@ def poly_mul_y_plus_q_inplace(P: np.ndarray, q: np.ndarray) -> None:
 
     # add q_part (which corresponds to q * old_P)
     P += q_part
+    
 
-
-def compute_P(
-        r: np.array, q_M: np.array, start_index, end_index
+def get_neighbors_shap_from_polynomials_archive(
+        M_f_i: np.array, R_i: float, q_M_left: np.array, q_M_right: np.array, f_w: np.array, left_leaf_weight: float, right_leaf_weight: float
 ):
-    """
-    Compute the polynomial (q_1 + y)*(q_2 + y)*..*(q_k + y).
-    """
-    P = np.zeros((len(r), q_M.shape[1]))
-    P[0, :] = np.prod(r)
-    for i in range(start_index, end_index):
-        poly_mul_y_plus_q_inplace(P, q_M[i])
+    M_left = M_f_i.copy()
+    poly_mul_y_plus_q_inplace_archive(M_left, q_M_left)
+    left_game_theory_metric_vector = (M_left * f_w).sum(axis=0) * left_leaf_weight * R_i
 
-    return P
+    M_right = M_f_i.copy()
+    poly_mul_y_plus_q_inplace_archive(M_right, q_M_right)
+    right_game_theory_metric_vector = (M_right * f_w).sum(axis=0) * right_leaf_weight * (1 - R_i)
 
+    return left_game_theory_metric_vector, right_game_theory_metric_vector
 
-def continue_P_compute(P: np.array, q_M: np.array, start_index: int, end_index: int):
+def continue_P_compute_archive(P: np.array, q_M: np.array, start_index: int, end_index: int):
     # assume P is NOT zfill
     P_continued = P.copy()
     for i in range(start_index, end_index):
-        poly_mul_y_plus_q_inplace(P_continued, q_M[i])
+        poly_mul_y_plus_q_inplace_archive(P_continued, q_M[i])
     return P_continued
 
- # compute_contribution_vectors_from_initial_P(P, r, q_M, f_w, w, start_index, end_index)
-def compute_contribution_vectors_from_initial_P(
+def compute_contribution_vectors_from_initial_P_archive(
         P: np.array, q_M: np.array, f_w: np.array, w: float, start_index: int, end_index: int
 ):
     """
@@ -106,148 +65,13 @@ def compute_contribution_vectors_from_initial_P(
             if j == i:
                 P_shared = M_f_i.copy()
                 continue
-            poly_mul_y_plus_q_inplace(M_f_i, q_M[j])
+            poly_mul_y_plus_q_inplace_archive(M_f_i, q_M[j])
 
         # Compute Shapley/Banzhaf values using the constructed polynomial
         game_theory_metric_vector = (M_f_i * f_w).sum(axis=0) * w
         constitutions_vectors.append(game_theory_metric_vector)
 
     return np.array(constitutions_vectors) # Now M become a |n| columns and |r| rows matrix
-
-
-def recursive_linear_tree_shap(P: np.array, r: np.array, q_M: np.array, f_w: np.array, w: float, start_index: int, end_index: int) -> np.array:
-    middle_index = (start_index + end_index) // 2
-    if end_index - start_index <= 4:
-        return compute_contribution_vectors_from_initial_P(P, q_M, f_w, w, start_index, end_index)
-
-    top_P = continue_P_compute(P, q_M, start_index=start_index, end_index=middle_index)
-    bottom_P = continue_P_compute(P, q_M, start_index=middle_index, end_index=end_index)
-
-    top_contribs = recursive_linear_tree_shap(bottom_P, r, q_M, f_w, w, start_index=start_index, end_index=middle_index)
-    bottom_contribs = recursive_linear_tree_shap(top_P, r, q_M, f_w, w, start_index=middle_index, end_index=end_index)
-    return np.vstack([top_contribs, bottom_contribs])
-
-
-def improved_linear_tree_shap_magic(r: np.array, p: np.array, f_w: np.array, w: float):
-    q_M = bits_matrix(p, len(r)) * (1 / r.reshape(-1, 1))
-    if len(r) <= 6:
-        initial_P = compute_P(r, q_M, start_index=0, end_index=0)
-        M = compute_contribution_vectors_from_initial_P(initial_P, q_M, f_w, w, start_index=0, end_index=len(r))
-        return (M * (q_M - 1)).T.copy()
-
-    top_P = compute_P(r, q_M, start_index=0, end_index=len(r)//2)
-    bottom_P = compute_P(r, q_M, start_index=len(r)//2, end_index=len(r))
-
-    bottom_contribs = recursive_linear_tree_shap(top_P, r, q_M, f_w, w, start_index=len(r)//2, end_index=len(r))
-    top_contribs = recursive_linear_tree_shap(bottom_P, r, q_M, f_w, w, start_index=0, end_index=len(r)//2)
-    M = np.vstack([top_contribs, bottom_contribs])
-    return (M * (q_M - 1)).T.copy()
-
-
-############################################################################################################################################################
-#
-#                         The Recursive Approach that Supports The Right Leaf Neighbor Trick
-#
-############################################################################################################################################################
-
-
-def get_neighbors_shap_from_polynomials(
-        M_f_i: np.array, R_i: float, q_M_left: np.array, q_M_right: np.array, f_w: np.array, left_leaf_weight: float, right_leaf_weight: float
-):
-    M_left = M_f_i.copy()
-    poly_mul_y_plus_q_inplace(M_left, q_M_left)
-    left_game_theory_metric_vector = (M_left * f_w).sum(axis=0) * left_leaf_weight * R_i
-
-    M_right = M_f_i.copy()
-    poly_mul_y_plus_q_inplace(M_right, q_M_right)
-    right_game_theory_metric_vector = (M_right * f_w).sum(axis=0) * right_leaf_weight * (1 - R_i)
-
-    return left_game_theory_metric_vector, right_game_theory_metric_vector
-
-def compute_contribution_vectors_from_initial_P_for_neighbors(
-        P: np.array, q_M: np.array, b_M: np.array, f_w: np.array, left_leaf_weight: float, right_leaf_weight: float, R_last: float, start_index: int, end_index: int
-):
-    """
-    Return a matrix with the Shapley/Banzhaf values contributions. The matrix rows are decision patterns (with the same
-    over as in the input p) and the columns are features contributions of the path features
-    (in the same order as the features cover appear in the input r)
-    """
-    # Longer, but numerically stable
-
-    last_row_q_M_left = q_M[-1]
-    last_row_q_M_right = (-(b_M[-1]-1)) * (1/(1-R_last)) # replace 0s and 1s and multiply by 1/(1-R_last)
-    left_constitutions_vectors = []
-    right_constitutions_vectors = []
-    P_shared = P
-    for i in range(start_index, end_index):
-        M_f_i = P_shared.copy()
-        for j in range(max(i-1, start_index), end_index):
-            if j == i:
-                P_shared = M_f_i.copy()
-                continue
-            poly_mul_y_plus_q_inplace(M_f_i, q_M[j])
-
-        # Compute Shapley/Banzhaf values using the constructed polynomial
-        left_game_theory_metric_vector, right_game_theory_metric_vector = get_neighbors_shap_from_polynomials(
-            M_f_i, R_last, last_row_q_M_left, last_row_q_M_right, f_w, left_leaf_weight, right_leaf_weight
-        )
-        left_constitutions_vectors.append(left_game_theory_metric_vector)
-        right_constitutions_vectors.append(right_game_theory_metric_vector)
-
-    if end_index + 1 == len(f_w):
-        poly_mul_y_plus_q_inplace(P_shared, q_M[-2])
-        left_constitutions_vectors.append((P_shared * f_w).sum(axis=0) * left_leaf_weight * R_last)
-        right_constitutions_vectors.append((P_shared * f_w).sum(axis=0) * right_leaf_weight * (1-R_last))
-
-    M_left = np.array(left_constitutions_vectors) # Now M become a |n| columns and |r| rows matrix
-    M_right = np.array(right_constitutions_vectors)
-    return M_left, M_right
-
-
-def recursive_linear_tree_shap_for_neighbors(
-        P: np.array, r: np.array, q_M: np.array, b_M: np.array, f_w: np.array, left_leaf_weight: float, right_leaf_weight: float, start_index: int, end_index: int
-) -> np.array:
-    middle_index = (start_index + end_index) // 2
-    if end_index - start_index <= 4:
-        R_last = r[-1]
-        return compute_contribution_vectors_from_initial_P_for_neighbors(P, q_M, b_M, f_w, left_leaf_weight, right_leaf_weight, R_last, start_index, end_index)
-
-    top_P = continue_P_compute(P, q_M, start_index=start_index, end_index=middle_index)
-    bottom_P = continue_P_compute(P, q_M, start_index=middle_index, end_index=end_index)
-
-    top_contribs_left, top_contribs_right = recursive_linear_tree_shap_for_neighbors(bottom_P, r, q_M, b_M, f_w, left_leaf_weight, right_leaf_weight, start_index=start_index, end_index=middle_index)
-    bottom_contribs_left, bottom_contribs_right = recursive_linear_tree_shap_for_neighbors(top_P, r, q_M, b_M, f_w, left_leaf_weight, right_leaf_weight, start_index=middle_index, end_index=end_index)
-    return np.vstack([top_contribs_left, bottom_contribs_left]), np.vstack([top_contribs_right, bottom_contribs_right]),
-
-
-def improved_linear_tree_shap_magic_for_neighbors(r: np.array, p: np.array, f_w: np.array, left_leaf_weight: float, right_leaf_weight: float):
-    b_M = bits_matrix(p, len(r))
-    neg_b_M = neg_bits_matrix(p, len(r))
-    q_M = b_M * (1 / r.reshape(-1, 1))
-
-    M_general = np.zeros((len(r), len(p)))
-    M_general[0, :] = np.prod(r[:-1])
-    if len(r) <= 6:
-        R_last = r[-1]
-        M_left, M_right = compute_contribution_vectors_from_initial_P_for_neighbors(
-            M_general, q_M, b_M, f_w, left_leaf_weight, right_leaf_weight, R_last, start_index=0, end_index=len(r)-1
-        )
-    else:
-        top_P = continue_P_compute(M_general, q_M, start_index=0, end_index=(len(r)-1)//2)
-        bottom_P = continue_P_compute(M_general, q_M, start_index=(len(r)-1)//2, end_index=len(r)-1)
-
-        bottom_contribs_left, bottom_contribs_right = recursive_linear_tree_shap_for_neighbors(top_P, r, q_M, b_M, f_w, left_leaf_weight, right_leaf_weight, start_index=(len(r)-1)//2, end_index=len(r)-1)
-        top_contribs_left, top_contribs_right = recursive_linear_tree_shap_for_neighbors(bottom_P, r, q_M, b_M, f_w, left_leaf_weight, right_leaf_weight, start_index=0, end_index=(len(r)-1)//2)
-        M_left = np.vstack([top_contribs_left, bottom_contribs_left])
-        M_right = np.vstack([top_contribs_right, bottom_contribs_right])
-
-    q_M_left = q_M
-    result_left = (M_left * (q_M_left - 1)).T.copy()
-    q_M_right = q_M.copy()
-    last_row_q_M_right = neg_b_M[-1] * (1 / (1 - r[-1]))
-    q_M_right[-1]=last_row_q_M_right
-    result_right = (M_right * (q_M_right - 1)).T.copy()
-    return result_left + result_right
 
 
 ############################################################################################################################################################
@@ -355,16 +179,16 @@ def linear_tree_shap_magic_for_neighbors(
             if j == i:
                 M_shared = M_f_i.copy()
                 continue
-            poly_mul_y_plus_q_inplace(M_f_i, q_M[j])
+            poly_mul_y_plus_q_inplace_archive(M_f_i, q_M[j])
 
         # Compute Shapley/Banzhaf values using the constructed polynomial
-        left_game_theory_metric_vector, right_game_theory_metric_vector = get_neighbors_shap_from_polynomials(
+        left_game_theory_metric_vector, right_game_theory_metric_vector = get_neighbors_shap_from_polynomials_archive(
             M_f_i, r[-1], last_row_q_M_left, last_row_q_M_right, f_w, left_leaf_weight, right_leaf_weight
         )
         left_constitutions_vectors.append(left_game_theory_metric_vector)
         right_constitutions_vectors.append(right_game_theory_metric_vector)
 
-    poly_mul_y_plus_q_inplace(M_shared, q_M[len(r)-2])
+    poly_mul_y_plus_q_inplace_archive(M_shared, q_M[len(r)-2])
     left_constitutions_vectors.append((M_shared * f_w).sum(axis=0) * left_leaf_weight * R_last)
     right_constitutions_vectors.append((M_shared * f_w).sum(axis=0) * right_leaf_weight * (1-R_last))
 
@@ -385,38 +209,6 @@ def linear_tree_shap_magic_for_neighbors(
 #
 ############################################################################################################################################################
 
-def extract_bit(patterns: np.ndarray, i: int):
-    """
-    patterns: np.array of integers
-    i: bit index to extract (0 = LSB)
-
-    Returns: patterns with bit i removed
-    """
-    lower = patterns & ((1 << i) - 1)     # bits below i
-    upper = patterns >> (i + 1)           # bits above i
-    return lower + (upper << i)
-
-def improved_linear_tree_shap_iv(r: np.array, p: np.array, f_w: np.array, w: float):
-    q_M = bits_matrix(p, len(r)) * (1 / r.reshape(-1, 1))
-    assert len(f_w) == len(r) - 1
-    shaps = []
-    for i, ratio in enumerate(r):
-        extracted_patterns = extract_bit(p, len(r) - 1 - i)
-        new_r = np.concatenate([r[:i], r[i+1:]])
-        shap_excluding_i = improved_linear_tree_shap_magic(new_r, extracted_patterns, f_w, w)
-        q_i = (np.tile(q_M[i], (q_M.shape[0] - 1, 1))).T
-        # The interaction values (i,j) are the shapley values of j in the game excluding i, times '(q_i - 1) * ratio'
-        # The '/ 2' is to fit the shap package logic, splitting the interaction between the two features.
-        shaps_i = (shap_excluding_i * (q_i - 1) * ratio / 2).copy()
-        shaps.append(shaps_i)
-    return shaps
-
-############################################################################################################################################################
-#
-#         Division Forward - Do not use it as it slower than the "improved" approach above. Keep it here as it might be useful someday
-#
-############################################################################################################################################################
-
 
 def linear_tree_shap_division_forward(
         r: np.array, p: np.array, f_w: np.array, leaf_weight: float
@@ -430,7 +222,7 @@ def linear_tree_shap_division_forward(
 
     M_general = np.zeros((len(r)+1, len(p)))
     M_general[0, :] = np.prod(r)
-    M_general = continue_P_compute(M_general, q_M, start_index=0, end_index=len(r))
+    M_general = continue_P_compute_archive(M_general, q_M, start_index=0, end_index=len(r))
 
     # Now M_general include the polynomials (y+q_0)*(y+q_1)*...*(y+q_k)
 
@@ -467,7 +259,7 @@ def linear_tree_shap_division_forward_for_neighbors(
 
     M_general = np.zeros((len(r), len(p)))
     M_general[0, :] = np.prod(r[:-1])
-    M_general = continue_P_compute(M_general, q_M, start_index=0, end_index=len(r) - 1)
+    M_general = continue_P_compute_archive(M_general, q_M, start_index=0, end_index=len(r) - 1)
 
     # Now M_general include the polynomials (y+q_0)*(y+q_1)*...*(y+q_k)
 
@@ -482,7 +274,7 @@ def linear_tree_shap_division_forward_for_neighbors(
             M_f_i[d] = (M_general[d] - M_f_i[d-1]) * one_over_q_M[i] + M_general[d+1] * neg_b_M[i]
 
         # Compute Shapley/Banzhaf values using the constructed polynomial
-        left_game_theory_metric_vector, right_game_theory_metric_vector = get_neighbors_shap_from_polynomials(
+        left_game_theory_metric_vector, right_game_theory_metric_vector = get_neighbors_shap_from_polynomials_archive(
             M_f_i, R_last, last_row_q_M_left, last_row_q_M_right, f_w, left_leaf_weight, right_leaf_weight
         )
         left_constitutions_vectors.append(left_game_theory_metric_vector)
@@ -516,7 +308,7 @@ QUAD_NODES   = DTYPE(0.5) * (_nodes_15.astype(DTYPE) + DTYPE(1.0))   # (n_quad,)
 QUAD_WEIGHTS = DTYPE(0.5) * _weights_15.astype(DTYPE)                 # (n_quad,)
 
 
-def linear_tree_shap_v6(
+def quadrature_tree_shap_batched_approach(
     r: np.ndarray,
     p: np.ndarray,
     leaf_value: float = 1.0,
@@ -584,7 +376,7 @@ def linear_tree_shap_v6(
     return phi
 
 
-def linear_tree_shap_v6_for_neighbors(
+def quadrature_tree_shap_batched_approach_for_neighbors(
     r: np.ndarray,
     decision_patterns: np.ndarray,
     w1: float = 1.0,
@@ -761,11 +553,11 @@ def linear_tree_shap_magic_blocked_draft(
     # Precompute: multiply all factors in the top half once, and all in bottom half once
     P_top_half = base_poly.copy()
     for j in top_idx:
-        poly_mul_y_plus_q_inplace(P_top_half, q_M[j])
+        poly_mul_y_plus_q_inplace_archive(P_top_half, q_M[j])
 
     P_bot_half = base_poly.copy()
     for j in bot_idx:
-        poly_mul_y_plus_q_inplace(P_bot_half, q_M[j])
+        poly_mul_y_plus_q_inplace_archive(P_bot_half, q_M[j])
 
     # For each block g compute P_out[g] = product over (y + q_j) for j not in block
     for block_index, block in enumerate(blocks):
@@ -782,7 +574,7 @@ def linear_tree_shap_magic_blocked_draft(
             to_compute_for_global_p = np.nonzero(~in_block)[0]
 
         for j in to_compute_for_global_p:
-            poly_mul_y_plus_q_inplace(P_shared, q_M[j])
+            poly_mul_y_plus_q_inplace_archive(P_shared, q_M[j])
 
         # For each i in this block, finish the local multiplications excluding i
         for i in range(block[0], block[-1]+1):
@@ -792,7 +584,7 @@ def linear_tree_shap_magic_blocked_draft(
                 if j == i:
                     P_shared = P.copy()
                     continue
-                poly_mul_y_plus_q_inplace(P, q_M[j])
+                poly_mul_y_plus_q_inplace_archive(P, q_M[j])
 
             # game_theory_metric_vector = sum_t P[t,:] * f_w[t]
             metric = (P * f_w).sum(axis=0)  # shape (n,)
@@ -1231,8 +1023,8 @@ def linear_tree_shap_magic_blocked(
     blocks = [list(range(start, min(start + block_size, k))) for start in range(0, k, block_size)]
 
     mid_index=blocks[G // 2][0]
-    P_top_half = continue_P_compute(P, q_M, start_index=0, end_index=mid_index)
-    P_bot_half = continue_P_compute(P, q_M, start_index=mid_index, end_index=k)
+    P_top_half = continue_P_compute_archive(P, q_M, start_index=0, end_index=mid_index)
+    P_bot_half = continue_P_compute_archive(P, q_M, start_index=mid_index, end_index=k)
 
     # For each block g compute P_out[g] = product over (y + q_j) for j not in block
     contribs = []
@@ -1245,8 +1037,8 @@ def linear_tree_shap_magic_blocked(
             to_compute_for_global_p = [i for i in range(mid_index, k) if i not in block]
 
         for j in to_compute_for_global_p:
-            poly_mul_y_plus_q_inplace(P_shared, q_M[j])
+            poly_mul_y_plus_q_inplace_archive(P_shared, q_M[j])
 
-        contribs.append(compute_contribution_vectors_from_initial_P(P_shared, q_M, f_w, leaf_weight, start_index=block[0], end_index=block[-1] + 1))
+        contribs.append(compute_contribution_vectors_from_initial_P_archive(P_shared, q_M, f_w, leaf_weight, start_index=block[0], end_index=block[-1] + 1))
 
     return (np.vstack(contribs) * (q_M - 1)).T.copy()
