+++ woodelf/lts_vectorized.py
@@ -1,135 +1,20 @@
-import math
-import time
-from math import factorial
-from typing import List, Any, Dict, Optional
+from typing import List, Any, Dict
 
 import numpy as np
 import pandas as pd
 from tqdm import tqdm
 
 from woodelf.core.cube_metric import CubeMetric, ShapleyInteractionValues, ShapleyValues, BanzhafValues
-from woodelf.core.trees.decision_trees_ensemble import DecisionTreeNode
 from woodelf.core.decision_patterns import decision_patterns_generator, ignore_right_neighbor
-from woodelf.core.lts_polynomial_multiplication import (
-    improved_linear_tree_shap_magic, improved_linear_tree_shap_magic_for_neighbors,
-    linear_tree_shap_magic_for_banzhaf, linear_tree_shap_division_forward, linear_tree_shap_magic, improved_linear_tree_shap_iv, linear_tree_shap_v6, linear_tree_shap_v6_for_neighbors
-)
-from woodelf.core.path_to_s_vectors.base_p2s import PathToSVectors
+from woodelf.core.path_to_s_vectors.lts_recursive_p2s import LTSRecursivePathToSVectors
+from woodelf.core.trees.decision_trees_ensemble import DecisionTreeNode
 from woodelf.core.trees.parse_models import load_decision_tree_ensemble_model
-
-
-def nCk(n, k):
-    return factorial(n) // (factorial(k) * factorial(n-k))
-
-def shapley_values_f_w(depth):
-    return np.array([[1 / (depth * nCk(depth-1, s))] for s in range(depth)])
-
-def banzhaf_values_f_w(depth):
-    return np.array([[1 / 2 ** (depth - 1)] for s in range(depth)])
-
-class LTSPathToSVectors(PathToSVectors):
-
-    def __init__(self, metric: CubeMetric, max_depth: int, GPU: bool = False):
-        super().__init__(metric, max_depth, GPU)
-
-        self.f_ws = None
-        if isinstance(metric, ShapleyInteractionValues) or isinstance(metric, ShapleyValues):
-            # None f_ws for depth 0, in the rest use shapley_values_f_w(depth)
-            self.f_ws = [None] + [shapley_values_f_w(depth) for depth in range(1, max_depth + 1)]
-
-        self.computation_time = 0
-
-    def _get_s_matrix(self, covers: np.array, consumer_patterns: np.array, w: float, w_neighbor: Optional[float] = None):
-        start_time = time.time()
-        # assume features in path are unique
-        if isinstance(self.metric, ShapleyInteractionValues):
-            assert w_neighbor is None
-            if len(covers) <= 1:
-                self.computation_time += time.time() - start_time
-                return []
-            f_w = self.f_ws[len(covers)-1]
-            s_matrix = improved_linear_tree_shap_iv(covers, consumer_patterns, f_w, w)
-        elif isinstance(self.metric, ShapleyValues):
-            f_w = self.f_ws[len(covers)]
-            if w_neighbor is None:
-                # s_matrix = linear_tree_shap_magic_faster_v2(covers, consumer_patterns, f_w, w)
-                s_matrix = improved_linear_tree_shap_magic(covers, consumer_patterns, f_w, w)
-            else:
-                s_matrix = improved_linear_tree_shap_magic_for_neighbors(covers, consumer_patterns, f_w, w, w_neighbor)
-        elif isinstance(self.metric, BanzhafValues):
-            if w_neighbor is not None:
-                s_matrix_left = linear_tree_shap_magic_for_banzhaf(covers, consumer_patterns, w)
-                covers_of_right = np.array(list(covers[:-1]) + [1 - covers[-1]])
-                consumer_patterns_right = consumer_patterns.copy()
-                consumer_patterns_right[consumer_patterns % 2 == 0] += 1
-                consumer_patterns_right[consumer_patterns % 2 == 1] -= 1
-                s_matrix_right = linear_tree_shap_magic_for_banzhaf(covers_of_right, consumer_patterns_right.astype(np.uint64), w_neighbor)
-                s_matrix = s_matrix_left + s_matrix_right
-            else:
-                s_matrix = linear_tree_shap_magic_for_banzhaf(covers, consumer_patterns, w)
-        else:
-            raise ValueError(f"Unsupported metric {self.metric.__class__}. The our LinearTreeSHAP implementation currently support only Shapley values, Banzhaf values and Shapley interaction values")
-        self.computation_time += time.time() - start_time
-        return s_matrix
-
-    def get_background_s_matrix(
-        self, features_in_path: List, consumer_patterns: np.ndarray,
-        background_patterns: np.ndarray, w: float, w_neighbor: Optional[float] = None
-    ) -> Dict:
-        raise NotImplementedError("LTS uses path-dependent approach only")
-
-    def get_path_dependent_s_matrix(
-        self, features_in_path: List, consumer_patterns: np.ndarray,
-        covers: np.ndarray, w: float, w_neighbor: Optional[float] = None
-    ) -> Dict:
-        """consumer_patterns must be unique (pre-factorized by the caller)."""
-        if not features_in_path:
-            return {}
-        s_matrix = self._get_s_matrix(covers, consumer_patterns, w, w_neighbor)
-        if not isinstance(self.metric, ShapleyInteractionValues):
-            s_matrix = s_matrix.astype(np.float32)
-            # TODO why np indexing on a matrix is slower than vector by vector! contribution_values = s_matrix[inverse]
-            return {feature: s_matrix[:, i] for i, feature in enumerate(features_in_path)}
-        else:
-            if not s_matrix:  # < 2 unique features in path, no interactions
-                return {}
-            result = {}
-            for i, feature1 in enumerate(features_in_path):
-                s_matrix_i = s_matrix[i].astype(np.float32)
-                for j, feature2 in enumerate(features_in_path):
-                    if i != j:
-                        f2_index = j if j < i else j - 1
-                        result[(feature1, feature2)] = s_matrix_i[:, f2_index]
-            return result
-
-    def present_statistics(self):
-        print(f"LTSPathToSVectors took {round(self.computation_time, 2)}")
-
-
-def get_unique_features_in_path(path: List[DecisionTreeNode]):
-    unique_features_in_path = []
-    for n in path:
-        if n.feature_name not in unique_features_in_path:
-            unique_features_in_path.append(n.feature_name)
-    return unique_features_in_path
-
-
-def get_covers_vector(path: List[DecisionTreeNode], unique_features_in_path: List[Any]):
-    if len(unique_features_in_path) == 0:
-        # If the leaf is the tree's root, it has cover of 1.
-        return [1]
-
-    feature_index = {f: i for i, f in enumerate(unique_features_in_path)}
-
-    proceed_covers = [1] * len(unique_features_in_path)
-    for i in range(len(path)-1):
-        proceed_covers[ feature_index[path[i].feature_name] ] *= (path[i+1].cover / path[i].cover)
-    return proceed_covers
+from woodelf.core.utils import get_unique_features_in_path, get_covers_vector
 
 
 def vectorized_linear_tree_shap_for_a_single_tree(
         tree: DecisionTreeNode, consumer_data: pd.DataFrame, values: Dict,
-        p2s: LTSPathToSVectors, GPU: bool, use_neighbor_leaf_trick: bool
+        p2s: LTSRecursivePathToSVectors, GPU: bool, use_neighbor_leaf_trick: bool
 ):
     leaf_index_to_covers = {}
     leaf_index_to_unique_features_in_path = {}
@@ -167,7 +52,7 @@ def vectorized_linear_tree_shap(
         model = load_decision_tree_ensemble_model(model, list(consumer_data.columns))
 
     if p2s_class is None:
-        p2s = LTSPathToSVectors(metric, max_depth=model.max_depth, GPU=GPU)
+        p2s = LTSRecursivePathToSVectors(metric, max_depth=model.max_depth, GPU=GPU)
     else:
         p2s = p2s_class(metric, max_depth=model.max_depth, GPU=GPU)
     values = {}
@@ -177,105 +62,3 @@ def vectorized_linear_tree_shap(
         )
     p2s.present_statistics()
     return values
-
-
-
-
-############################################################################################################################################################
-#
-#                         For Tests and Measures
-#
-############################################################################################################################################################
-
-
-class LTSSimpleNeighborTrickPathToSVectors(LTSPathToSVectors):
-
-    def poly_mult_shap_func(self, covers: np.array, consumer_patterns: np.array, f_w: np.array, w: float):
-        raise NotImplemented()
-
-    def poly_mult_banzhaf_func(self, covers: np.array, consumer_patterns: np.array, w: float):
-        return linear_tree_shap_magic_for_banzhaf(covers, consumer_patterns, w)
-
-    def _get_s_matrix(self, covers: np.array, consumer_patterns: np.array, w: float, w_neighbor: Optional[float] = None):
-        start_time = time.time()
-        if isinstance(self.metric, ShapleyValues):
-            # assume features in path are unique
-            f_w = self.f_ws[len(covers)]
-            if w_neighbor is None:
-                s_matrix = self.poly_mult_shap_func(covers, consumer_patterns, f_w, w)
-            else:
-                s_matrix_left = self.poly_mult_shap_func(covers, consumer_patterns, f_w, w)
-                covers_of_right = np.array(list(covers[:-1]) + [1 - covers[-1]])
-                consumer_patterns_right = consumer_patterns.copy()
-                consumer_patterns_right[consumer_patterns % 2 == 0] += 1
-                consumer_patterns_right[consumer_patterns % 2 == 1] -= 1
-                s_matrix_right = self.poly_mult_shap_func(
-                    covers_of_right, consumer_patterns_right.astype(np.uint64), f_w, w_neighbor
-                )
-                s_matrix = s_matrix_left + s_matrix_right
-        else:
-            if w_neighbor is None:
-                s_matrix = self.poly_mult_banzhaf_func(covers, consumer_patterns, w)
-            else:
-                s_matrix_left = self.poly_mult_banzhaf_func(covers, consumer_patterns, w)
-                covers_of_right = np.array(list(covers[:-1]) + [1 - covers[-1]])
-                consumer_patterns_right = consumer_patterns.copy()
-                consumer_patterns_right[consumer_patterns % 2 == 0] += 1
-                consumer_patterns_right[consumer_patterns % 2 == 1] -= 1
-                s_matrix_right = self.poly_mult_banzhaf_func(covers_of_right, consumer_patterns_right.astype(np.uint64), w_neighbor)
-                s_matrix = s_matrix_left + s_matrix_right
-        self.computation_time += time.time() - start_time
-        return s_matrix
-
-class LTSSimplePathToSVectors(LTSSimpleNeighborTrickPathToSVectors):
-
-    def poly_mult_shap_func(self, covers: np.array, consumer_patterns: np.array, f_w: np.array, w: float):
-        return linear_tree_shap_magic(covers, consumer_patterns, f_w, w)
-
-
-class LTSImprovedPathToSVectors(LTSSimpleNeighborTrickPathToSVectors):
-
-    def poly_mult_shap_func(self, covers: np.array, consumer_patterns: np.array, f_w: np.array, w: float):
-        if len(covers) <= 36:
-            return linear_tree_shap_division_forward(covers, consumer_patterns, f_w, w)
-        return improved_linear_tree_shap_magic(covers, consumer_patterns, f_w, w)
-
-
-class LTSV6PathToSVectors(LTSPathToSVectors):
-
-    def __init__(self, metric: CubeMetric, max_depth: int, GPU: bool = False):
-        super().__init__(metric, max_depth, GPU)
-        self.quad_nodes, self.quad_weights = self.compute_quads()
-
-    def compute_quads(self):
-        max_required_quads = min(max(int(math.ceil(self.max_depth / 2)), 2), 16)
-        quad_nodes = {}
-        quad_weights = {}
-        for n_quad in range(2, max_required_quads + 1):
-            _nodes, _weights = np.polynomial.legendre.leggauss(n_quad)
-            quad_nodes[n_quad] = np.float64(0.5) * (_nodes.astype(np.float64) + np.float64(1.0))  # (n_quad,)
-            quad_weights[n_quad] = np.float64(0.5) * _weights.astype(np.float64)  # (n_quad,)
-        return quad_nodes, quad_weights
-
-    def _get_s_matrix(self, covers: np.array, consumer_patterns: np.array, w: float, w_neighbor: Optional[float] = None):
-        start_time = time.time()
-        if isinstance(self.metric, ShapleyValues):
-            # For D<4 use 2, for D > 32 use 16 for 4<=D<=32 use 0.5*D
-            n_quads = min(max(int(math.ceil(len(covers) / 2)), 2), 16)
-            if w_neighbor is None:
-                s_matrix = linear_tree_shap_v6(covers, consumer_patterns, w, self.quad_nodes[n_quads], self.quad_weights[n_quads])
-            else:
-                s_matrix = linear_tree_shap_v6_for_neighbors(covers, consumer_patterns, w, w_neighbor, self.quad_nodes[n_quads], self.quad_weights[n_quads])
-        else:
-            if w_neighbor is None:
-                s_matrix = linear_tree_shap_magic_for_banzhaf(covers, consumer_patterns, w)
-            else:
-                s_matrix_left = linear_tree_shap_magic_for_banzhaf(covers, consumer_patterns, w)
-                covers_of_right = np.array(list(covers[:-1]) + [1 - covers[-1]])
-                consumer_patterns_right = consumer_patterns.copy()
-                consumer_patterns_right[consumer_patterns % 2 == 0] += 1
-                consumer_patterns_right[consumer_patterns % 2 == 1] -= 1
-                s_matrix_right = linear_tree_shap_magic_for_banzhaf(covers_of_right, consumer_patterns_right.astype(np.int64), w_neighbor)
-                s_matrix = s_matrix_left + s_matrix_right
-        self.computation_time += time.time() - start_time
-        return s_matrix
